Update generateId : (int -> string) to use the 'len' parameter.
Just a random typo I noticed while reading. Not a huge change. 'List.replicate' was hardcoded to use 5. I was browsing random F# projects and saw this issue. Small typos tend to be the worst to find and the compiler doesn't show any warnings about it, so I thought I'd be friendly and point it out.